1. Fixed bug in counsellor chat page list (loading image issue 2. overall call will be disposed if we refreshed the page
Remaining issue: when call is made, and canceled before picking up, it's still in other person's UI even though it's getting deleted in real time db, as overlay is not getting removed. (need to check)
video call cutting issue: if a person who is picking the call will cut the call, he is getting redirected to blank page. issue is with context, somehow it is getting context of overlay, and not the actual context